refactor: move webview IPC command constants to src/shared/ipc/
The seven *Commands.ts files in src/common/webview/ were pure string-constant
objects defining the IPC protocol between webview frontends and the extension
host.  They had zero dependencies yet lived in the "VS Code-allowed zone",
which CLAUDE.md requires to be kept separate from platform-agnostic code.

This caused two documented layer violations:
- src/controllers/ (VS Code-free zone) imported from @common/webview/
- src/shared/settingsView/handlers/ (webview frontend) imported from
  the VS Code-allowed zone for string constants

Fix: move all seven constants files to src/shared/ipc/, which is already
accessible via @shared/* in every package's tsconfig.  The originals in
src/common/webview/ become one-line re-export shims so all existing consumers
(desktop, extension, tests) continue to work without changes.  The seven
violating imports in src/controllers/ and src/shared/settingsView/handlers/
are updated to import from @shared/ipc/ directly.
