+++ src/shared/ipc/commonCommands.ts
@@ -1,0 +1,8 @@
+export const COMMON_COMMANDS = {
+  THEME_SET: 'setTheme',
+  DEBUG_MODE_SET: 'setDebugMode',
+  STATE_RESTORE: 'restoreState',
+  WEBVIEW_READY: 'webviewReady',
+  ERROR: 'error',
+  SWITCH_VIEW: 'switchView',
+} as const;
