+++ src/shared/ipc/index.ts
@@ -1,0 +1,7 @@
+export { COMMON_COMMANDS } from './commonCommands';
+export { MAIN_VIEW_COMMANDS } from './mainViewCommands';
+export { PROGRESS_VIEW_COMMANDS } from './progressViewCommands';
+export { HISTORY_VIEW_COMMANDS } from './historyViewCommands';
+export { PROFILE_VIEW_COMMANDS } from './profileViewCommands';
+export { MEMORY_VIEW_COMMANDS } from './memoryViewCommands';
+export { SETTINGS_VIEW_CMD, SETTINGS_VIEW_COMMANDS } from './settingsViewCommands';
