+++ src/shared/ipc/progressViewCommands.ts
@@ -1,0 +1,87 @@
+import { COMMON_COMMANDS } from './commonCommands';
+
+export const PROGRESS_VIEW_COMMANDS = {
+  ...COMMON_COMMANDS,
+
+  SWITCH_STREAM: 'switchStream',
+  DELETE_STREAM: 'deleteStream',
+  CLEAN_STREAM: 'cleanStream',
+  STOP_STREAM: 'stopStream',
+  UPDATE_STREAMS: 'updateStreams',
+  DELETE_ALL: 'deleteAll',
+
+  LOG_DELTA: 'logDelta',
+
+  UPDATE_TODOS: 'updateTodos',
+  UPDATE_PLAN: 'updatePlan',
+  UPDATE_QUEUED_FOLLOW_UPS: 'updateQueuedFollowUps',
+  SYNC_STREAM_CONTENT: 'syncStreamContent',
+
+  UPDATE_STREAM_STATUS: 'updateStreamStatus',
+  SET_ACTIVE_STREAM: 'setActiveStream',
+  UPDATE_CONVERSATION_PROGRESS: 'updateConversationProgress',
+  UPDATE_STREAM_BADGES: 'updateStreamBadges',
+  UPDATE_STREAM_DESCRIPTION: 'updateStreamDescription',
+  UPDATE_PROCESS_OUTPUT: 'updateProcessOutput',
+  SYNC_INQUIRY_THREADS: 'syncInquiryThreads',
+  UPDATE_INQUIRY_THREAD: 'updateInquiryThread',
+  UPDATE_PARENT_STREAM: 'updateParentStream',
+  UPDATE_FILES: 'updateFiles',
+  UPDATE_MISSING_OUTPUTS: 'updateMissingOutputs',
+  UPDATE_COMPILE_FAILURES: 'updateCompileFailures',
+  UPDATE_PERMISSION: 'updatePermission',
+  UPDATE_BYPASS: 'updateBypass',
+
+  UPDATE_RUN_USAGE: 'updateRunUsage',
+
+  RESUME: 'resume',
+  RUN_NEW: 'runNew',
+  COMPACT_RESPONSE: 'compactResponse',
+  RETRY_STREAM_REQUEST: 'retryStreamRequest',
+  CANCEL_RETRY_REQUEST: 'cancelRetryRequest',
+  USE_OWN_API_KEY: 'useOwnApiKey',
+  DIFF_STREAM: 'diffStream',
+  PACK_STREAM: 'packStream',
+  FILTER_STREAMS: 'filterStreams',
+  RESTORE_STATE: 'restoreState',
+  SEND_FOLLOW_UP: 'sendFollowUp',
+  POLISH_FOLLOW_UP: 'polishFollowUp',
+  UPDATE_FOLLOW_UP_TEXT: 'updateFollowUpText',
+  SETUP_FOLLOWUP: 'setupFollowup',
+  RUN_FOLLOWUP: 'runFollowup',
+  GET_FOLLOWUP_OPTIONS: 'getFollowupOptions',
+  SET_FOLLOWUP_OPTIONS: 'setFollowupOptions',
+  START_RECORDING: 'startRecording',
+  STOP_RECORDING: 'stopRecording',
+  UPDATE_RECORDING: 'updateRecording',
+  SHOW_INFORMATION_MESSAGE: 'showInformationMessage',
+  OPEN_TASK_STORAGE: 'openTaskStorage',
+  RUN_COMPILE_FIXER: 'runCompileFixer',
+  TOOL_EDIT_APPROVAL_ACTION: 'toolEditApprovalAction',
+  TOGGLE_TOOL_EDIT_APPROVAL_BYPASS: 'toggleToolEditApprovalBypass',
+  AGENT_PROPOSAL_ACTION: 'agentProposalAction',
+  BASH_APPROVAL_ACTION: 'bashApprovalAction',
+  PLAN_APPROVAL_ACTION: 'planApprovalAction',
+  EXTERNAL_INQUIRY_ACTION: 'externalInquiryAction',
+  USER_QUESTION_ACTION: 'userQuestionAction',
+  RESTORE_PROPOSAL_CONFIG: 'restoreProposalConfig',
+  TOGGLE_SUPER_YOLO_BYPASS: 'toggleSuperYoloBypass',
+  ODYSSEY_ACTIVE_UPDATED: 'odysseyActiveUpdated',
+
+  OPEN_MEMORY_VIEW: 'openMemoryView',
+
+  OPEN_FILE: 'openFile',
+  OPEN_FILE_COMPILE: 'openFileCompile',
+  COMPARE_ORIGINAL: 'compareOriginal',
+  COMPARE_PREVIOUS: 'comparePrevious',
+  ACCEPT_FILE: 'acceptFile',
+  MERGE_FILE: 'mergeFile',
+  LATEXDIFF_FILE: 'latexdiffFile',
+  OPEN_LABEL: 'openLabel',
+
+  OPEN_PROFILE: 'openProfile',
+
+  POP_OUT: 'popOut',
+  POP_BACK: 'popBack',
+  SET_PLACEMENT: 'setPlacement',
+} as const;
