+++ src/shared/settingsView/handlers/superYoloHandlers.ts
@@ -7,7 +7,7 @@
  * extension; absent in the desktop build), so callers supply them.
  */
 import { WorkspaceStateKey } from '@common/state/stateKeys';
-import { SETTINGS_VIEW_COMMANDS } from '@common/webview/settingsViewCommands';
+import { SETTINGS_VIEW_COMMANDS } from '@shared/ipc/settingsViewCommands';
 import {
   NESTED_DELEGATION_DEPTH_RANGE,
   clampNestedDelegationDepth,
